api/requirements.txt
Flask
PyMySQL
python-dotenv
Cerberus